requirements.txt
gradio>=4.0,<5.0
huggingface_hub<1.0.0
fastapi>=0.111
uvicorn[standard]
anthropic>=0.25
python-dotenv
